Fix Command_SetTeam and Frame_AlternateJoinTeam
Command_SetTeam would reply with success even though it failed as a result of you using an incorrect argument for the team. Appropriately set the result to state that it failed.

Frame_AlternateJoinTeam would place the GOTV bot on a team and spawn them. Added a check to exclude GOTV.